fix manual langauge selection
